Add routing to app
