+++ package.json
@@ -18,6 +18,8 @@
     "@types/react-dom": "16.9.4",
     "@types/react-leaflet": "^2.5.0",
     "@types/uuid": "^8.0.0",
+    "@types/react-router": "^5.1.7",
+    "@types/react-router-dom": "^5.1.5",
     "clsx": "^1.1.0",
     "debug": "^4.1.1",
     "enzyme": "^3.11.0",
@@ -30,6 +32,8 @@
     "react-dom": "^16.12.0",
     "react-leaflet": "^2.6.0",
     "react-leaflet-control": "^2.1.1",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "^3.4.0",
     "serve": "^11.3.2",
     "sweetalert2": "^9.13.1",
