+++ src/components/login/loginForm.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const LoginForm = function() {
+  return <div>test</div>;
+};
+
+export default LoginForm;
