+++ src/components/login/registerForm.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const RegisterForm = function() {
+  return <div>register</div>;
+};
+
+export default RegisterForm;
