+++ src/util/url.ts
@@ -1,0 +1,4 @@
+export const DASHBOARD_ROUTE = '/dashboard';
+export const LOGIN_ROUTE = '/login';
+export const REGISTER_ROUTE = '/register/';
+export const USER_ACCOUNT_ROUTE = '/account';
